Specify requirement that pylibmc be in requirements.txt
Instructing that a requirement should be added to a project's
requirements.txt file is not, on its face, taken to mean that a
requirement must be added to that file specifically. pip allows for
including files so it suffices that a requirement be added to an
included requirements file. In this case however the pylibmc requirement
must be explicitly added to the requirements.txt file. That should be
stated explicitly with explanation.